Change pin labelling to match IDC connector pinout
When assigning the default 2x05 IDC 2.54mm connector footprint to this symbol, the pins get swapped around so that the 'red stripe' would get connected to +12V, not -12V. In every Eurorack power cable I've looked at, pin 1 (indicated by the subtle arrow on the outside of the connector) aligns with the red stripe and the -12V rail. It's a subtle error, and doesn't matter if you're using unshrouded pin headers and keep track of the pin numbering yourself, but it's very important and potentially hard to catch if you're using the IDC connectors.

(Please test this change before integrating. I made the change by hand in GitHub after verifying and fixing the problem on my local machine, so I can't guarantee I haven't made a typo.)